LayoutTests/css3/transform-value-types.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../fast/js/resources/js-test-style.css">
<script src="../fast/js/resources/js-test-pre.js"></script>
<style id="style" media="screen">
.non-existent-class {
    -webkit-transform: translate(10px)
                       translateX(10px)
                       translateY(10px)
                       rotate(10deg)
                       scale(2)
                       scaleX(1.2)    /* test uppercase */
                       scaley(1.2)
                       skew(10deg,      10rad)
                       skewX(10deg)
                       skewy(11grad)
                       matrix(1, 0, 0,     1, 0, 0);
}
</style>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script type="text/javascript" charset="utf-8">

  var stylesheet = document.getElementById("style").sheet;
  var root = document.documentElement;

  function jsWrapperClass(node)
  {
      if (!node)
          return "[null]";
      var string = Object.prototype.toString.apply(node);
      return string.substr(8, string.length - 9);
  }

  function test(expression, className, prototypeName, constructorName)
  {
      if (!prototypeName)
          prototypeName = className + "Prototype";
      if (!constructorName)
          constructorName = className + "Constructor";
      shouldBe("jsWrapperClass(" + expression + ")", "'" + className + "'");
      shouldBe("jsWrapperClass(" + expression + ".__proto__)", "'" + prototypeName + "'");
      shouldBe("jsWrapperClass(" + expression + ".constructor)", "'" + constructorName + "'");
  }

  var transformRule = stylesheet.cssRules.item(0).style.getPropertyCSSValue('-webkit-transform');
  test("transformRule", "CSSValueList");
  
  test("transformRule[0]", "WebKitCSSTransformValue");

  shouldBe("transformRule[0].operationType", "1");        // CSS_TRANSLATE
  shouldBe("transformRule[0].cssText", "'translate(10px)'");

  shouldBe("transformRule[1].operationType", "2");        // CSS_TRANSLATEX
  shouldBe("transformRule[1].cssText", "'translateX(10px)'");

  shouldBe("transformRule[2].operationType", "3");        // CSS_TRANSLATEY
  shouldBe("transformRule[2].cssText", "'translateY(10px)'");

  shouldBe("transformRule[3].operationType", "4");        // CSS_ROTATE
  shouldBe("transformRule[3].cssText", "'rotate(10deg)'");

  shouldBe("transformRule[4].operationType", "5");        // CSS_SCALE
  shouldBe("transformRule[4].cssText", "'scale(2)'");

  shouldBe("transformRule[5].operationType", "6");        // CSS_SCALEX
  shouldBe("transformRule[5].cssText", "'scaleX(1.2)'");

  shouldBe("transformRule[6].operationType", "7");        // CSS_SCALEY
  shouldBe("transformRule[6].cssText", "'scaleY(1.2)'");

  shouldBe("transformRule[7].operationType", "8");        // CSS_SKEW
  shouldBe("transformRule[7].cssText", "'skew(10deg, 10rad)'");

  shouldBe("transformRule[8].operationType", "9");        // CSS_SKEWX
  shouldBe("transformRule[8].cssText", "'skewX(10deg)'");

  shouldBe("transformRule[9].operationType", "10");        // CSS_SKEWY
  shouldBe("transformRule[9].cssText", "'skewY(11grad)'");

  shouldBe("transformRule[10].operationType", "11");        // CSS_MATRIX
  shouldBe("transformRule[10].cssText", "'matrix(1, 0, 0, 1, 0, 0)'");
  
  var successfullyParsed = true;
</script>
<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
