LayoutTests/platform/chromium-win/css3/transform-value-types-expected.txt

PASS jsWrapperClass(transformRule) is 'CSSValueList'
FAIL jsWrapperClass(transformRule.__proto__) should be CSSValueListPrototype. Was Object.
FAIL jsWrapperClass(transformRule.constructor) should be CSSValueListConstructor. Was Function.
PASS jsWrapperClass(transformRule[0]) is 'WebKitCSSTransformValue'
FAIL jsWrapperClass(transformRule[0].__proto__) should be WebKitCSSTransformValuePrototype. Was Object.
FAIL jsWrapperClass(transformRule[0].constructor) should be WebKitCSSTransformValueConstructor. Was Function.
PASS transformRule[0].operationType is 1
PASS transformRule[0].cssText is 'translate(10px)'
PASS transformRule[1].operationType is 2
PASS transformRule[1].cssText is 'translateX(10px)'
PASS transformRule[2].operationType is 3
PASS transformRule[2].cssText is 'translateY(10px)'
PASS transformRule[3].operationType is 4
PASS transformRule[3].cssText is 'rotate(10deg)'
PASS transformRule[4].operationType is 5
PASS transformRule[4].cssText is 'scale(2)'
PASS transformRule[5].operationType is 6
PASS transformRule[5].cssText is 'scaleX(1.2)'
PASS transformRule[6].operationType is 7
PASS transformRule[6].cssText is 'scaleY(1.2)'
PASS transformRule[7].operationType is 8
PASS transformRule[7].cssText is 'skew(10deg, 10rad)'
PASS transformRule[8].operationType is 9
PASS transformRule[8].cssText is 'skewX(10deg)'
PASS transformRule[9].operationType is 10
PASS transformRule[9].cssText is 'skewY(11grad)'
PASS transformRule[10].operationType is 11
PASS transformRule[10].cssText is 'matrix(1, 0, 0, 1, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE
